+++ communication/src/main/java/com/appcoins/communication/sender/MessageSender.java
@@ -1,4 +1,4 @@
-package com.appcoins.communication;
+package com.appcoins.communication.sender;
 
 import android.content.Context;
 import android.content.Intent;
